requirements.txt
langchain==0.2.12
sentence-transformers==3.0.1
torch==2.3.1
pypdf==4.3.1
python-dotenv==1.0.1
qdrant-client==1.11.1
langchain-community==0.2.11
langchain-groq==0.1.9
Flask==3.0.3
langchain_qdrant==0.1.3
fastembed==0.2.4
openai==1.3.6
langchainhub==0.1.14
matplotlib==3.9.2
psycopg2-binary
spacy
tensorflow
tf-keras==2.17.0
fpdf
requests 
beautifulsoup4
lxml